+++ src/Fp.Tests/Processor_Debug.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Linq;
 using System.Text;
 using Fp.Tests.Utility;
 using NUnit.Framework;
@@ -16,6 +17,11 @@ public class Processor_Debug : ProcessorTestBase
         var results = P.MemLabel(data, pattern);
         Assert.That(results.Count, Is.EqualTo(2));
         Assert.That(P.MemAnnotations.Count, Is.EqualTo(1));
+        var kv = P.MemAnnotations.First();
+        Assert.That(kv.Key, Is.EqualTo((ReadOnlyMemory<byte>)data.AsMemory()));
+        Assert.That(kv.Value.Count, Is.EqualTo(2));
+        Assert.That(kv.Value.Values[0].Offset, Is.EqualTo(4));
+        Assert.That(kv.Value.Values[1].Offset, Is.EqualTo(7));
     }
 
     [Test]
@@ -26,6 +32,24 @@ public class Processor_Debug : ProcessorTestBase
         var results = P.MemLabel(data, pattern);
         Assert.That(results.Count, Is.EqualTo(2));
         Assert.That(P.MemAnnotations.Count, Is.EqualTo(0));
+    }
+
+    [Test]
+    public void MemLabel_RepeatingSequence_HasConcatenated()
+    {
+        P.Debug = true;
+        byte[] data = Encoding.UTF8.GetBytes("super_ababab_drome_abababab");
+        Span<byte> pattern = stackalloc byte[] { (byte)'a', (byte)'b' };
+        var results = P.MemLabel(data, pattern);
+        Assert.That(results.Count, Is.EqualTo(2));
+        Assert.That(P.MemAnnotations.Count, Is.EqualTo(1));
+        var kv = P.MemAnnotations.First();
+        Assert.That(kv.Key, Is.EqualTo((ReadOnlyMemory<byte>)data.AsMemory()));
+        Assert.That(kv.Value.Count, Is.EqualTo(2));
+        Assert.That(kv.Value.Values[0].Offset, Is.EqualTo(6));
+        Assert.That(kv.Value.Values[0].Length, Is.EqualTo(6));
+        Assert.That(kv.Value.Values[1].Offset, Is.EqualTo(19));
+        Assert.That(kv.Value.Values[1].Length, Is.EqualTo(8));
     }
 
     // TODO